Create README.md
The database is an organized collection of structured data to make it easily accessible, manageable, and update. In simple words, you can say, a database is a place where the data is stored. The best analogy is the library. 
The library contains a huge collection of books of different genres, here the library is a database and books are the data.